+++ client/src/App.tsx
@@ -1,9 +1,9 @@
 import "./App.css";
 import styled from "styled-components";
 import { Route, Routes, BrowserRouter } from "react-router-dom";
-import Filter from "./_components/Filter";
+import Filter from "./_components/Filter/Filter";
 import { DataProvider } from "./providers/DataProvider";
-import CardContainer from "./_components/CardContainer";
+import CardContainer from "./_components/Card/CardContainer";
 import { FilterProvider } from "./providers/FilterProvider";
 import { ReqUrlProvider } from "./providers/ReqUrlProvider";
