+++ client/src/_components/Filter/Filter.style.ts
@@ -1,0 +1,19 @@
+import styled from "styled-components";
+
+export const FilterContainer = styled.div`
+  display: flex;
+  gap: 5px;
+`;
+
+export const RefreshButton = styled.div`
+  width: 20px;
+  height: 20px;
+  border: 1px solid #ccc;
+  border-radius: 9999px;
+  background-color: #f2f2f2;
+  cursor: pointer;
+  padding: 6px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
