+++ client/src/_components/Filter/FilterItem.tsx
@@ -1,25 +1,6 @@
-import styled from "styled-components";
 import { memo } from "react";
 
-const FilterBox = styled.div`
-  position: relative;
-`;
-
-const FilterOption = styled.div`
-  border: 1px solid #ccc;
-  border-radius: 6px;
-  padding: 6px 10px;
-  background-color: white;
-  cursor: pointer;
-  font-size: 14px;
-  white-space: nowrap;
-  max-width: 200px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  &:hover {
-    background-color: #f9f9f9;
-  }
-`;
+import { FilterBox, FilterOption } from "./FilterItem.style";
 
 type Props = {
   label: string;
