+++ client/src/type.ts
@@ -1,0 +1,29 @@
+export type BankProduct = {
+  id: number;
+  companyName: string;
+  companyCode: string;
+  interestRate: string;
+  primeInterestRate: string;
+  depositAmount: number;
+  name: string;
+};
+
+type StateType = { isActive: boolean; text: string };
+
+export type FilterStateType = {
+  bank: StateType;
+  rate: StateType;
+  deposit: StateType;
+};
+
+export type ReqUrlType = {
+  companyCode: string;
+  interestRate: string;
+  deposit: string;
+};
+
+export type BankList = {
+  name: string;
+  companyCode: string;
+  checked: boolean;
+};
